Force flutter to be built
Currently, to get flutter to be rebuilt you must delete the snapshot or
create a commit. With this change it can be rebuilt by setting the
BUILD_TOOL environment variable. Adding BUIlD_TOOL will always force
flutter to be rebuilt when used with any command. If all you want to do
is to rebuild flutter the noop command was added.

Typical usage:

 $ BUILD_TOOL=1 flutter run

Or using noop:

 $ BUILD_TOOL=1 flutter noop